Enhance booking model and controller: add booking and payment status, update appointment creation logic, and improve data handling in booking history

+++ lib/model/session_type_model.dart
@@ -7,7 +7,7 @@ class SessionTypeModel {
   String name = "";
   String description= "";
   String duration= "";
-  double price = 0;
+  int price = 0;
   String? imageUrl= null;
 
   SessionTypeModel({
